Rename Adapter and Middleware TestCase
