+++ test/tesla/adapter/hackney_test.exs
@@ -1,8 +1,8 @@
 defmodule Tesla.Adapter.HackneyTest do
   use ExUnit.Case
-  use Tesla.Adapter.TestCase.Basic, adapter: :hackney
-  use Tesla.Adapter.TestCase.StreamRequestBody, adapter: :hackney
-  use Tesla.Adapter.TestCase.SSL, adapter: :hackney
+  use Tesla.AdapterCase.Basic, adapter: :hackney
+  use Tesla.AdapterCase.StreamRequestBody, adapter: :hackney
+  use Tesla.AdapterCase.SSL, adapter: :hackney
 
   test "get with `with_body: true` option" do
     defmodule Client do
@@ -11,7 +11,7 @@ defmodule Tesla.Adapter.HackneyTest do
       adapter :hackney, with_body: true
     end
 
-    response = Client.get("#{Tesla.Adapter.TestCase.http_url()}/ip")
+    response = Client.get("#{Tesla.AdapterCase.http_url()}/ip")
     assert response.status == 200
   end
 end
