+++ test/tesla/adapter/ibrowse_test.exs
@@ -1,6 +1,6 @@
 defmodule Tesla.Adapter.IbrowseTest do
   use ExUnit.Case
-  use Tesla.Adapter.TestCase.Basic, adapter: :ibrowse
-  use Tesla.Adapter.TestCase.StreamRequestBody, adapter: :ibrowse
-  use Tesla.Adapter.TestCase.SSL, adapter: :ibrowse
+  use Tesla.AdapterCase.Basic, adapter: :ibrowse
+  use Tesla.AdapterCase.StreamRequestBody, adapter: :ibrowse
+  use Tesla.AdapterCase.SSL, adapter: :ibrowse
 end
