+++ test/tesla/middleware/form_urlencoded_test.exs
@@ -1,7 +1,7 @@
 defmodule FormUrlencodedTest do
   use ExUnit.Case
 
-  use Tesla.Middleware.TestCase, middleware: Tesla.Middleware.FormUrlencoded
+  use Tesla.MiddlewareCase, middleware: Tesla.Middleware.FormUrlencoded
 
   defmodule Client do
     use Tesla
