+++ test/tesla/middleware/logger_test.exs
@@ -1,8 +1,8 @@
 defmodule LoggerTest do
   use ExUnit.Case, async: false
 
-  use Tesla.Middleware.TestCase, middleware: Tesla.Middleware.Logger
-  use Tesla.Middleware.TestCase, middleware: Tesla.Middleware.DebugLogger
+  use Tesla.MiddlewareCase, middleware: Tesla.Middleware.Logger
+  use Tesla.MiddlewareCase, middleware: Tesla.Middleware.DebugLogger
 
   defmodule Client do
     use Tesla
